+++ src/rpc/mod.rs
@@ -20,18 +20,8 @@ mod wallet_api;
 use std::sync::Arc;
 
 use crate::rpc_api::{
-    auth_api::*,
-    beacon_api::*,
-    chain_api::*,
-    common_api::*,
-    data_types::RPCState,
-    eth_api::{ETH_ACCOUNTS, ETH_BLOCK_NUMBER, ETH_CHAIN_ID},
-    gas_api::*,
-    mpool_api::*,
-    net_api::*,
-    node_api::NODE_STATUS,
-    state_api::*,
-    sync_api::*,
+    auth_api::*, beacon_api::*, chain_api::*, common_api::*, data_types::RPCState, eth_api::*,
+    gas_api::*, mpool_api::*, net_api::*, node_api::NODE_STATUS, state_api::*, sync_api::*,
     wallet_api::*,
 };
 use axum::routing::{get, post};
@@ -130,6 +120,7 @@ where
             .with_method(STATE_MARKET_BALANCE, state_market_balance::<DB>)
             .with_method(STATE_MARKET_DEALS, state_market_deals::<DB>)
             .with_method(STATE_MINER_INFO, state_miner_info::<DB>)
+            .with_method(MINER_GET_BASE_INFO, miner_get_base_info::<DB>)
             .with_method(STATE_MINER_ACTIVE_SECTORS, state_miner_active_sectors::<DB>)
             .with_method(STATE_MINER_FAULTS, state_miner_faults::<DB>)
             .with_method(STATE_MINER_RECOVERIES, state_miner_recoveries::<DB>)
@@ -179,6 +170,7 @@ where
             .with_method(ETH_ACCOUNTS, eth_api::eth_accounts)
             .with_method(ETH_BLOCK_NUMBER, eth_api::eth_block_number::<DB>)
             .with_method(ETH_CHAIN_ID, eth_api::eth_chain_id::<DB>)
+            .with_method(ETH_GAS_PRICE, eth_api::eth_gas_price::<DB>)
             .finish_unwrapped(),
     );