+++ src/rpc_api/mod.rs
@@ -82,6 +82,7 @@ pub static ACCESS_MAP: Lazy<HashMap<&str, Access>> = Lazy::new(|| {
     access.insert(state_api::STATE_MARKET_BALANCE, Access::Read);
     access.insert(state_api::STATE_MARKET_DEALS, Access::Read);
     access.insert(state_api::STATE_MINER_INFO, Access::Read);
+    access.insert(state_api::MINER_GET_BASE_INFO, Access::Read);
     access.insert(state_api::STATE_MINER_ACTIVE_SECTORS, Access::Read);
     access.insert(state_api::STATE_MINER_FAULTS, Access::Read);
     access.insert(state_api::STATE_MINER_RECOVERIES, Access::Read);
@@ -131,6 +132,7 @@ pub static ACCESS_MAP: Lazy<HashMap<&str, Access>> = Lazy::new(|| {
     access.insert(eth_api::ETH_ACCOUNTS, Access::Read);
     access.insert(eth_api::ETH_BLOCK_NUMBER, Access::Read);
     access.insert(eth_api::ETH_CHAIN_ID, Access::Read);
+    access.insert(eth_api::ETH_GAS_PRICE, Access::Read);
 
     access
 });
@@ -255,6 +257,7 @@ pub mod state_api {
     pub const STATE_MARKET_BALANCE: &str = "Filecoin.StateMarketBalance";
     pub const STATE_MARKET_DEALS: &str = "Filecoin.StateMarketDeals";
     pub const STATE_MINER_INFO: &str = "Filecoin.StateMinerInfo";
+    pub const MINER_GET_BASE_INFO: &str = "Filecoin.MinerGetBaseInfo";
     pub const STATE_MINER_FAULTS: &str = "Filecoin.StateMinerFaults";
     pub const STATE_MINER_RECOVERIES: &str = "Filecoin.StateMinerRecoveries";
     pub const STATE_MINER_POWER: &str = "Filecoin.StateMinerPower";
@@ -375,7 +378,33 @@ pub mod node_api {
 
 // Eth API
 pub mod eth_api {
+    use num_bigint::BigInt;
+    use serde::{Deserialize, Serialize};
+
+    use crate::lotus_json::lotus_json_with_self;
+
     pub const ETH_ACCOUNTS: &str = "Filecoin.EthAccounts";
     pub const ETH_BLOCK_NUMBER: &str = "Filecoin.EthBlockNumber";
     pub const ETH_CHAIN_ID: &str = "Filecoin.EthChainId";
+    pub const ETH_GAS_PRICE: &str = "Filecoin.EthGasPrice";
+
+    #[derive(Debug, Deserialize, Serialize, Default)]
+    pub struct GasPriceResult(#[serde(with = "crate::lotus_json::hexify")] pub BigInt);
+
+    lotus_json_with_self!(GasPriceResult);
+
+    #[cfg(test)]
+    mod test {
+        use super::*;
+        use quickcheck_macros::quickcheck;
+
+        #[quickcheck]
+        fn gas_price_result_serde_roundtrip(i: u128) {
+            let r = GasPriceResult(i.into());
+            let encoded = serde_json::to_string(&r).unwrap();
+            assert_eq!(encoded, format!("\"0x{i:x}\""));
+            let decoded: GasPriceResult = serde_json::from_str(&encoded).unwrap();
+            assert_eq!(r.0, decoded.0);
+        }
+    }
 }